^ Version Bump. $ Reverted Utils.java (Fixes ArrayList<ArrayList> cannot be converted to List<Object>) $ Fixed use of generics in MultiPickaxeBase.java.

+++ src/Java/gtPlusPlus/core/util/item/ItemUtils.java
@@ -311,7 +311,7 @@ public class ItemUtils {
 
 	public static MultiPickaxeBase generateMultiPick(final boolean GT_Durability, final Materials material){
 		final ToolMaterial customMaterial = Utils.generateToolMaterialFromGT(material);
-		final Object enchant = new Pair(material.mEnchantmentTools, material.mEnchantmentToolsLevel);
+		final Object enchant = new Pair(material.mEnchantmentTools, (Byte.valueOf(material.mEnchantmentToolsLevel)));
 		return generateMultiPick(GT_Durability, customMaterial, material.mDefaultLocalName, material.mDurability, material.mRGBa, enchant);
 	}
